Extract splice sites from annotation

This function extracts splice sites from an annotation object (a gtf/gff3 file, a GRanges object or a TxDb object) and saves them in a text file formatted such that it can be directly used with HISAT2, by providing it as the argument known-splicesite-infile.

extract_splice_sites(features, outfile, min_length = 5)

Arguments

features
Either the path to a gtf/gff3 file containing the genomic features, a GRanges object or a TxDb object.

outfile
Character scalar. The path to a text file where the extracted splice sites will be written.

min_length
Integer scalar. Junctions corresponding to introns below this size will not be reported. The default setting in HISAT2 is 5.

Value

Nothing is returned, but the splice junction coordinates are written to outfile.

References

Kim D, Langmead B and Salzberg SL. HISAT: a fast spliced aligner with low memory requirements. Nature Methods 12:357-360 (2015).

Author

Charlotte Soneson

Examples

tmp <- tempfile()
extract_splice_sites(features=system.file("extdata/refs/genes.gtf",
                                          package="Rhisat2"),
                     outfile=tmp, min_length=5)
#> Import genomic features from the file as a GRanges object ... 
#> OK
#> Prepare the 'metadata' data frame ... 
#> OK
#> Make the TxDb object ... 
#> OK