Align reads with HISAT2

The function can be used to call the hisat2 binary.

hisat2(
  sequences,
  index,
  ...,
  type = c("single", "paired"),
  outfile,
  force = FALSE,
  strict = TRUE,
  execute = TRUE
)

Arguments

sequences
If type is single, a character vector of file names if the additional argument c is FALSE, otherwise a vector of read sequences. If type is paired, a length-2 list of file names or sequences, where the first list item corresponds to the first mate pair sequences, and the second list item to the second mate pair sequences.

index
Character scalar. The path+prefix of the HISAT2 index to align against (in the form <path/to/index>/<prefix>).

...
Additional arguments passed to the binaries.

type
Character scalar, either "single" or "paired". If single, the input sequences are interpreted as single-end reads. If paired, they are supposed to be paired reads.

outfile
(optional) Character scalar. The path to the output file. If missing, the alignments will be returned as an R character vector.

force
Logical scalar. Whether to force overwriting of outdir.

strict
Logical scalar. Whether strict checking of input arguments should be enforced.

execute
Logical scalar. Whether to execute the assembled shell command. If FALSE, return a string with the command.

Value

If execute is TRUE, the output generated by calling the hisat2 binary. If execute is FALSE, the hisat2 command.

Details

All additional arguments in ... are interpreted as additional arguments to the HISAT2 binaries. Any flags are supposed to be represented as logical values (e.g., quiet=TRUE will be translated into --quiet). Parameters with additional input are supposed to be character or numeric vectors, and the individual elements are collapsed into a single comma-separated string (e.g., k=2 is translated into -k 2, bmax=100 into --bmax 100). Some arguments to the HISAT2 binaries will be ignored if they are already handled as explicit function arguments. See the output of hisat2_usage() for details about available parameters.

References

Kim D, Langmead B and Salzberg SL. HISAT: a fast spliced aligner with low memory requirements. Nature Methods 12:357-360 (2015).

Author

Charlotte Soneson, based on code from Florian Hahne.

Examples

tmp <- tempdir()
refs <- list.files(system.file("extdata/refs", package="Rhisat2"),
                   full.names=TRUE, pattern="\\.fa$")
hisat2_build(references=refs, outdir=file.path(tmp, "index"),
             force=TRUE, prefix="index")
reads <- list.files(system.file("extdata/reads", package="Rhisat2"),
                    full.names=TRUE, pattern="\\.fastq$")
hisat2(sequences=as.list(reads), index=file.path(tmp, "index/index"),
       type="paired", outfile=file.path(tmp, "out.sam"), force=TRUE)