Print usage of hisat2

Print usage of hisat2

hisat2_usage()

Value

No value is returned, the usage of hisat2 is printed to the console.

Author

Charlotte Soneson

Examples

hisat2_usage()
#>   [1] "HISAT2 version 2.1.0 by Daehwan Kim (infphilo@gmail.com, www.ccb.jhu.edu/people/infphilo)"                  
#>   [2] "Usage: "                                                                                                    
#>   [3] "  hisat2-align [options]* -x <ht2-idx> {-1 <m1> -2 <m2> | -U <r>} [-S <sam>]"                               
#>   [4] ""                                                                                                           
#>   [5] "  <ht2-idx>  Index filename prefix (minus trailing .X.ht2)."                                                
#>   [6] "  <m1>       Files with #1 mates, paired with files in <m2>."                                               
#>   [7] "  <m2>       Files with #2 mates, paired with files in <m1>."                                               
#>   [8] "  <r>        Files with unpaired reads."                                                                    
#>   [9] "  <sam>      File for SAM output (default: stdout)"                                                         
#>  [10] ""                                                                                                           
#>  [11] "  <m1>, <m2>, <r> can be comma-separated lists (no whitespace) and can be"                                  
#>  [12] "  specified many times.  E.g. '-U file1.fq,file2.fq -U file3.fq'."                                          
#>  [13] ""                                                                                                           
#>  [14] "Options (defaults in parentheses):"                                                                         
#>  [15] ""                                                                                                           
#>  [16] " Input:"                                                                                                    
#>  [17] "  -q                 query input files are FASTQ .fq/.fastq (default)"                                      
#>  [18] "  --qseq             query input files are in Illumina's qseq format"                                       
#>  [19] "  -f                 query input files are (multi-)FASTA .fa/.mfa"                                          
#>  [20] "  -r                 query input files are raw one-sequence-per-line"                                       
#>  [21] "  -c                 <m1>, <m2>, <r> are sequences themselves, not files"                                   
#>  [22] "  -s/--skip <int>    skip the first <int> reads/pairs in the input (none)"                                  
#>  [23] "  -u/--upto <int>    stop after first <int> reads/pairs (no limit)"                                         
#>  [24] "  -5/--trim5 <int>   trim <int> bases from 5'/left end of reads (0)"                                        
#>  [25] "  -3/--trim3 <int>   trim <int> bases from 3'/right end of reads (0)"                                       
#>  [26] "  --phred33          qualities are Phred+33 (default)"                                                      
#>  [27] "  --phred64          qualities are Phred+64"                                                                
#>  [28] "  --int-quals        qualities encoded as space-delimited integers"                                         
#>  [29] ""                                                                                                           
#>  [30] " Alignment:"                                                                                                
#>  [31] "  --n-ceil <func>    func for max # non-A/C/G/Ts permitted in aln (L,0,0.15)"                               
#>  [32] "  --ignore-quals     treat all quality values as 30 on Phred scale (off)"                                   
#>  [33] "  --nofw             do not align forward (original) version of read (off)"                                 
#>  [34] "  --norc             do not align reverse-complement version of read (off)"                                 
#>  [35] ""                                                                                                           
#>  [36] " Spliced Alignment:"                                                                                        
#>  [37] "  --pen-cansplice <int>              penalty for a canonical splice site (0)"                               
#>  [38] "  --pen-noncansplice <int>           penalty for a non-canonical splice site (12)"                          
#>  [39] "  --pen-canintronlen <func>          penalty for long introns (G,-8,1) with canonical splice sites"         
#>  [40] "  --pen-noncanintronlen <func>       penalty for long introns (G,-8,1) with noncanonical splice sites"      
#>  [41] "  --min-intronlen <int>              minimum intron length (20)"                                            
#>  [42] "  --max-intronlen <int>              maximum intron length (500000)"                                        
#>  [43] "  --known-splicesite-infile <path>   provide a list of known splice sites"                                  
#>  [44] "  --novel-splicesite-outfile <path>  report a list of splice sites"                                         
#>  [45] "  --novel-splicesite-infile <path>   provide a list of novel splice sites"                                  
#>  [46] "  --no-temp-splicesite               disable the use of splice sites found"                                 
#>  [47] "  --no-spliced-alignment             disable spliced alignment"                                             
#>  [48] "  --rna-strandness <string>          specify strand-specific information (unstranded)"                      
#>  [49] "  --tmo                              reports only those alignments within known transcriptome"              
#>  [50] "  --dta                              reports alignments tailored for transcript assemblers"                 
#>  [51] "  --dta-cufflinks                    reports alignments tailored specifically for cufflinks"                
#>  [52] "  --avoid-pseudogene                 tries to avoid aligning reads to pseudogenes (experimental option)\023"
#>  [53] "  --no-templatelen-adjustment        disables template length adjustment for RNA-seq reads"                 
#>  [54] ""                                                                                                           
#>  [55] " Scoring:"                                                                                                  
#>  [56] "  --mp <int>,<int>   max and min penalties for mismatch; lower qual = lower penalty <6,2>"                  
#>  [57] "  --sp <int>,<int>   max and min penalties for soft-clipping; lower qual = lower penalty <2,1>"             
#>  [58] "  --no-softclip      no soft-clipping"                                                                      
#>  [59] "  --np <int>         penalty for non-A/C/G/Ts in read/ref (1)"                                              
#>  [60] "  --rdg <int>,<int>  read gap open, extend penalties (5,3)"                                                 
#>  [61] "  --rfg <int>,<int>  reference gap open, extend penalties (5,3)"                                            
#>  [62] "  --score-min <func> min acceptable alignment score w/r/t read length"                                      
#>  [63] "                     (L,0.0,-0.2)"                                                                          
#>  [64] ""                                                                                                           
#>  [65] " Reporting:"                                                                                                
#>  [66] "  -k <int> (default: 5) report up to <int> alns per read"                                                   
#>  [67] ""                                                                                                           
#>  [68] " Paired-end:"                                                                                               
#>  [69] "  -I/--minins <int>  minimum fragment length (0), only valid with --no-spliced-alignment"                   
#>  [70] "  -X/--maxins <int>  maximum fragment length (500), only valid with --no-spliced-alignment"                 
#>  [71] "  --fr/--rf/--ff     -1, -2 mates align fw/rev, rev/fw, fw/fw (--fr)"                                       
#>  [72] "  --no-mixed         suppress unpaired alignments for paired reads"                                         
#>  [73] "  --no-discordant    suppress discordant alignments for paired reads"                                       
#>  [74] ""                                                                                                           
#>  [75] " Output:"                                                                                                   
#>  [76] "  -t/--time          print wall-clock time taken by search phases"                                          
#>  [77] "  --summary-file     print alignment summary to this file."                                                 
#>  [78] "  --new-summary      print alignment summary in a new style, which is more machine-friendly."               
#>  [79] "  --quiet            print nothing to stderr except serious errors"                                         
#>  [80] "  --met-file <path>  send metrics to file at <path> (off)"                                                  
#>  [81] "  --met-stderr       send metrics to stderr (off)"                                                          
#>  [82] "  --met <int>        report internal counters & metrics every <int> secs (1)"                               
#>  [83] "  --no-head          supppress header lines, i.e. lines starting with @"                                    
#>  [84] "  --no-sq            supppress @SQ header lines"                                                            
#>  [85] "  --rg-id <text>     set read group id, reflected in @RG line and RG:Z: opt field"                          
#>  [86] "  --rg <text>        add <text> (\"lab:value\") to @RG line of SAM header."                                 
#>  [87] "                     Note: @RG line only printed when --rg-id is set."                                      
#>  [88] "  --omit-sec-seq     put '*' in SEQ and QUAL fields for secondary alignments."                              
#>  [89] ""                                                                                                           
#>  [90] " Performance:"                                                                                              
#>  [91] "  -o/--offrate <int> override offrate of index; must be >= index's offrate"                                 
#>  [92] "  -p/--threads <int> number of alignment threads to launch (1)"                                             
#>  [93] "  --reorder          force SAM output order to match order of input reads"                                  
#>  [94] "  --mm               use memory-mapped I/O for index; many 'hisat2's can share"                             
#>  [95] ""                                                                                                           
#>  [96] " Other:"                                                                                                    
#>  [97] "  --qc-filter        filter out reads that are bad according to QSEQ filter"                                
#>  [98] "  --seed <int>       seed for random number generator (0)"                                                  
#>  [99] "  --non-deterministic seed rand. gen. arbitrarily instead of using read attributes"                         
#> [100] "  --remove-chrname   remove 'chr' from reference names in alignment"                                        
#> [101] "  --add-chrname      add 'chr' to reference names in alignment "                                            
#> [102] "  --version          print version information and quit"                                                    
#> [103] "  -h/--help          print this usage message"