Print usage of hisat2

Print usage of hisat2

hisat2_usage()

Value

No value is returned, the usage of hisat2 is printed to the console.

Author

Charlotte Soneson

Examples

hisat2_usage()
#>   [1] "HISAT2 version 2.2.1 by Daehwan Kim (infphilo@gmail.com, www.ccb.jhu.edu/people/infphilo)"                                                                                                         
#>   [2] "Usage: "                                                                                                                                                                                           
#>   [3] "  hisat2-align [options]* -x <ht2-idx> {-1 <m1> -2 <m2> | -U <r>} [-S <sam>]"                                                                                                                      
#>   [4] ""                                                                                                                                                                                                  
#>   [5] "  <ht2-idx>  Index filename prefix (minus trailing .X.ht2)."                                                                                                                                       
#>   [6] "  <m1>       Files with #1 mates, paired with files in <m2>."                                                                                                                                      
#>   [7] "  <m2>       Files with #2 mates, paired with files in <m1>."                                                                                                                                      
#>   [8] "  <r>        Files with unpaired reads."                                                                                                                                                           
#>   [9] "  <sam>      File for SAM output (default: stdout)"                                                                                                                                                
#>  [10] ""                                                                                                                                                                                                  
#>  [11] "  <m1>, <m2>, <r> can be comma-separated lists (no whitespace) and can be"                                                                                                                         
#>  [12] "  specified many times.  E.g. '-U file1.fq,file2.fq -U file3.fq'."                                                                                                                                 
#>  [13] ""                                                                                                                                                                                                  
#>  [14] "Options (defaults in parentheses):"                                                                                                                                                                
#>  [15] ""                                                                                                                                                                                                  
#>  [16] " Input:"                                                                                                                                                                                           
#>  [17] "  -q                 query input files are FASTQ .fq/.fastq (default)"                                                                                                                             
#>  [18] "  --qseq             query input files are in Illumina's qseq format"                                                                                                                              
#>  [19] "  -f                 query input files are (multi-)FASTA .fa/.mfa"                                                                                                                                 
#>  [20] "  -r                 query input files are raw one-sequence-per-line"                                                                                                                              
#>  [21] "  -c                 <m1>, <m2>, <r> are sequences themselves, not files"                                                                                                                          
#>  [22] "  -s/--skip <int>    skip the first <int> reads/pairs in the input (none)"                                                                                                                         
#>  [23] "  -u/--upto <int>    stop after first <int> reads/pairs (no limit)"                                                                                                                                
#>  [24] "  -5/--trim5 <int>   trim <int> bases from 5'/left end of reads (0)"                                                                                                                               
#>  [25] "  -3/--trim3 <int>   trim <int> bases from 3'/right end of reads (0)"                                                                                                                              
#>  [26] "  --phred33          qualities are Phred+33 (default)"                                                                                                                                             
#>  [27] "  --phred64          qualities are Phred+64"                                                                                                                                                       
#>  [28] "  --int-quals        qualities encoded as space-delimited integers"                                                                                                                                
#>  [29] ""                                                                                                                                                                                                  
#>  [30] " Presets:                 Same as:"                                                                                                                                                                
#>  [31] "   --fast                 --no-repeat-index"                                                                                                                                                       
#>  [32] "   --sensitive            --bowtie2-dp 1 -k 30 --score-min L,0,-0.5"                                                                                                                               
#>  [33] "   --very-sensitive       --bowtie2-dp 2 -k 50 --score-min L,0,-1"                                                                                                                                 
#>  [34] ""                                                                                                                                                                                                  
#>  [35] " Alignment:"                                                                                                                                                                                       
#>  [36] "  --bowtie2-dp <int> use Bowtie2's dynamic programming alignment algorithm (0) - 0: no dynamic programming, 1: conditional dynamic programming, and 2: unconditional dynamic programming (slowest)"
#>  [37] "  --n-ceil <func>    func for max # non-A/C/G/Ts permitted in aln (L,0,0.15)"                                                                                                                      
#>  [38] "  --ignore-quals     treat all quality values as 30 on Phred scale (off)"                                                                                                                          
#>  [39] "  --nofw             do not align forward (original) version of read (off)"                                                                                                                        
#>  [40] "  --norc             do not align reverse-complement version of read (off)"                                                                                                                        
#>  [41] "  --no-repeat-index  do not use repeat index"                                                                                                                                                      
#>  [42] ""                                                                                                                                                                                                  
#>  [43] " Spliced Alignment:"                                                                                                                                                                               
#>  [44] "  --pen-cansplice <int>              penalty for a canonical splice site (0)"                                                                                                                      
#>  [45] "  --pen-noncansplice <int>           penalty for a non-canonical splice site (12)"                                                                                                                 
#>  [46] "  --pen-canintronlen <func>          penalty for long introns (G,-8,1) with canonical splice sites"                                                                                                
#>  [47] "  --pen-noncanintronlen <func>       penalty for long introns (G,-8,1) with noncanonical splice sites"                                                                                             
#>  [48] "  --min-intronlen <int>              minimum intron length (20)"                                                                                                                                   
#>  [49] "  --max-intronlen <int>              maximum intron length (500000)"                                                                                                                               
#>  [50] "  --known-splicesite-infile <path>   provide a list of known splice sites"                                                                                                                         
#>  [51] "  --novel-splicesite-outfile <path>  report a list of splice sites"                                                                                                                                
#>  [52] "  --novel-splicesite-infile <path>   provide a list of novel splice sites"                                                                                                                         
#>  [53] "  --no-temp-splicesite               disable the use of splice sites found"                                                                                                                        
#>  [54] "  --no-spliced-alignment             disable spliced alignment"                                                                                                                                    
#>  [55] "  --rna-strandness <string>          specify strand-specific information (unstranded)"                                                                                                             
#>  [56] "  --tmo                              reports only those alignments within known transcriptome"                                                                                                     
#>  [57] "  --dta                              reports alignments tailored for transcript assemblers"                                                                                                        
#>  [58] "  --dta-cufflinks                    reports alignments tailored specifically for cufflinks"                                                                                                       
#>  [59] "  --avoid-pseudogene                 tries to avoid aligning reads to pseudogenes (experimental option)"                                                                                           
#>  [60] "  --no-templatelen-adjustment        disables template length adjustment for RNA-seq reads"                                                                                                        
#>  [61] ""                                                                                                                                                                                                  
#>  [62] " Scoring:"                                                                                                                                                                                         
#>  [63] "  --mp <int>,<int>   max and min penalties for mismatch; lower qual = lower penalty <6,2>"                                                                                                         
#>  [64] "  --sp <int>,<int>   max and min penalties for soft-clipping; lower qual = lower penalty <2,1>"                                                                                                    
#>  [65] "  --no-softclip      no soft-clipping"                                                                                                                                                             
#>  [66] "  --np <int>         penalty for non-A/C/G/Ts in read/ref (1)"                                                                                                                                     
#>  [67] "  --rdg <int>,<int>  read gap open, extend penalties (5,3)"                                                                                                                                        
#>  [68] "  --rfg <int>,<int>  reference gap open, extend penalties (5,3)"                                                                                                                                   
#>  [69] "  --score-min <func> min acceptable alignment score w/r/t read length"                                                                                                                             
#>  [70] "                     (L,0.0,-0.2)"                                                                                                                                                                 
#>  [71] ""                                                                                                                                                                                                  
#>  [72] " Reporting:"                                                                                                                                                                                       
#>  [73] "  -k <int>           It searches for at most <int> distinct, primary alignments for each read. Primary alignments mean "                                                                           
#>  [74] "                     alignments whose alignment score is equal to or higher than any other alignments. The search terminates "                                                                     
#>  [75] "                     when it cannot find more distinct valid alignments, or when it finds <int>, whichever happens first. "                                                                        
#>  [76] "                     The alignment score for a paired-end alignment equals the sum of the alignment scores of "                                                                                    
#>  [77] "                     the individual mates. Each reported read or pair alignment beyond the first has the SAM ‘secondary’ bit "                                                                     
#>  [78] "                     (which equals 256) set in its FLAGS field. For reads that have more than <int> distinct, "                                                                                    
#>  [79] "                     valid alignments, hisat2 does not guarantee that the <int> alignments reported are the best possible "                                                                        
#>  [80] "                     in terms of alignment score. Default: 5 (linear index) or 10 (graph index)."                                                                                                  
#>  [81] "                     Note: HISAT2 is not designed with large values for -k in mind, and when aligning reads to long, "                                                                             
#>  [82] "                     repetitive genomes, large -k could make alignment much slower."                                                                                                               
#>  [83] "  --max-seeds <int>  HISAT2, like other aligners, uses seed-and-extend approaches. HISAT2 tries to extend seeds to "                                                                               
#>  [84] "                     full-length alignments. In HISAT2, --max-seeds is used to control the maximum number of seeds that "                                                                          
#>  [85] "                     will be extended. For DNA-read alignment (--no-spliced-alignment), HISAT2 extends up to these many seeds"                                                                     
#>  [86] "                     and skips the rest of the seeds. For RNA-read alignment, HISAT2 skips extending seeds and reports "                                                                           
#>  [87] "                     no alignments if the number of seeds is larger than the number specified with the option, "                                                                                   
#>  [88] "                     to be compatible with previous versions of HISAT2. Large values for --max-seeds may improve alignment "                                                                       
#>  [89] "                     sensitivity, but HISAT2 is not designed with large values for --max-seeds in mind, and when aligning "                                                                        
#>  [90] "                     reads to long, repetitive genomes, large --max-seeds could make alignment much slower. "                                                                                      
#>  [91] "                     The default value is the maximum of 5 and the value that comes with -k times 2."                                                                                              
#>  [92] "  -a/--all           HISAT2 reports all alignments it can find. Using the option is equivalent to using both --max-seeds "                                                                         
#>  [93] "                     and -k with the maximum value that a 64-bit signed integer can represent (9,223,372,036,854,775,807)."                                                                        
#>  [94] "  --repeat           report alignments to repeat sequences directly"                                                                                                                               
#>  [95] ""                                                                                                                                                                                                  
#>  [96] " Paired-end:"                                                                                                                                                                                      
#>  [97] "  -I/--minins <int>  minimum fragment length (0), only valid with --no-spliced-alignment"                                                                                                          
#>  [98] "  -X/--maxins <int>  maximum fragment length (500), only valid with --no-spliced-alignment"                                                                                                        
#>  [99] "  --fr/--rf/--ff     -1, -2 mates align fw/rev, rev/fw, fw/fw (--fr)"                                                                                                                              
#> [100] "  --no-mixed         suppress unpaired alignments for paired reads"                                                                                                                                
#> [101] "  --no-discordant    suppress discordant alignments for paired reads"                                                                                                                              
#> [102] ""                                                                                                                                                                                                  
#> [103] " Output:"                                                                                                                                                                                          
#> [104] "  -t/--time          print wall-clock time taken by search phases"                                                                                                                                 
#> [105] "  --summary-file <path> print alignment summary to this file."                                                                                                                                     
#> [106] "  --new-summary         print alignment summary in a new style, which is more machine-friendly."                                                                                                   
#> [107] "  --quiet               print nothing to stderr except serious errors"                                                                                                                             
#> [108] "  --met-file <path>     send metrics to file at <path> (off)"                                                                                                                                      
#> [109] "  --met-stderr          send metrics to stderr (off)"                                                                                                                                              
#> [110] "  --met <int>           report internal counters & metrics every <int> secs (1)"                                                                                                                   
#> [111] "  --no-head             suppress header lines, i.e. lines starting with @"                                                                                                                         
#> [112] "  --no-sq               suppress @SQ header lines"                                                                                                                                                 
#> [113] "  --rg-id <text>        set read group id, reflected in @RG line and RG:Z: opt field"                                                                                                              
#> [114] "  --rg <text>           add <text> (\"lab:value\") to @RG line of SAM header."                                                                                                                     
#> [115] "                        Note: @RG line only printed when --rg-id is set."                                                                                                                          
#> [116] "  --omit-sec-seq        put '*' in SEQ and QUAL fields for secondary alignments."                                                                                                                  
#> [117] ""                                                                                                                                                                                                  
#> [118] " Performance:"                                                                                                                                                                                     
#> [119] "  -o/--offrate <int> override offrate of index; must be >= index's offrate"                                                                                                                        
#> [120] "  -p/--threads <int> number of alignment threads to launch (1)"                                                                                                                                    
#> [121] "  --reorder          force SAM output order to match order of input reads"                                                                                                                         
#> [122] "  --mm               use memory-mapped I/O for index; many 'hisat2's can share"                                                                                                                    
#> [123] ""                                                                                                                                                                                                  
#> [124] " Other:"                                                                                                                                                                                           
#> [125] "  --qc-filter        filter out reads that are bad according to QSEQ filter"                                                                                                                       
#> [126] "  --seed <int>       seed for random number generator (0)"                                                                                                                                         
#> [127] "  --non-deterministic seed rand. gen. arbitrarily instead of using read attributes"                                                                                                                
#> [128] "  --remove-chrname   remove 'chr' from reference names in alignment"                                                                                                                               
#> [129] "  --add-chrname      add 'chr' to reference names in alignment "                                                                                                                                   
#> [130] "  --version          print version information and quit"                                                                                                                                           
#> [131] "  -h/--help          print this usage message"